pyproject.toml
[project]
name = "paganini-aios"
version = "2.0.0"
description = "AI Operating System for Financial Markets"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "Apache-2.0"}
authors = [
    {name = "Rod Marques", email = "rod.marques@aios.finance"},
    {name = "João Raf"},
]

dependencies = [
    "click>=8.0",
    "rich>=13.0",
    "pyyaml>=6.0",
    "fastapi>=0.100.0",
    "uvicorn[standard]>=0.20.0",
    "litellm>=1.0",
    "chromadb>=1.0",
    "httpx>=0.27",
    "pdfplumber>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.4",
    "black>=24.0",
    "isort>=5.0",
]
rl = [
    "torch>=2.0",
    "transformers>=4.40",
]
all = ["paganini-aios[dev,rl]"]

[project.scripts]
paganini = "core.channels.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["core", "packs"]

[tool.ruff]
target-version = "py312"
line-length = 100
